Move files into /indoor and /outdoor
