+++ lib/models/indoor/indoor_location.dart
@@ -1,7 +1,7 @@
 import 'package:concordia_navigation/models/indoor/coordinate.dart';
 import 'package:concordia_navigation/models/indoor/floor.dart';
 import 'package:concordia_navigation/models/outdoor/building.dart';
-import 'package:concordia_navigation/models/reachable.dart';
+import 'package:concordia_navigation/models/outdoor/reachable.dart';
 import 'package:concordia_navigation/models/uni_location.dart';
 import 'package:concordia_navigation/services/search.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
