+++ lib/models/indoor/node.dart
