+++ lib/models/outdoor/outdoor_location.dart
@@ -1,4 +1,4 @@
-import 'package:concordia_navigation/models/reachable.dart';
+import 'package:concordia_navigation/models/outdoor/reachable.dart';
 import 'package:concordia_navigation/models/uni_location.dart';
 import 'package:concordia_navigation/services/search.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
