+++ lib/models/outdoor/reachable.dart
