+++ lib/models/outdoor/university.dart
