+++ lib/screens/indoor_page.dart
@@ -1,6 +1,6 @@
 import 'package:concordia_navigation/models/indoor/floor.dart';
 import 'package:concordia_navigation/providers/buildings_data.dart';
-import 'package:concordia_navigation/services/painters.dart';
+import 'package:concordia_navigation/services/indoor/painters.dart';
 import 'package:concordia_navigation/widgets/indoor/indoor_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_navigation/storage/app_constants.dart' as constants;
