+++ lib/services/outdoor/directions_service.dart
@@ -1,4 +1,4 @@
-import 'package:concordia_navigation/services/network.dart';
+import 'package:concordia_navigation/services/outdoor/network.dart';
 import 'package:concordia_navigation/storage/app_constants.dart' as constants;
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:meta/meta.dart' show visibleForTesting;
