+++ lib/services/outdoor/indoor_itinerary.dart
@@ -1,6 +1,6 @@
 import 'package:concordia_navigation/models/indoor/indoor_location.dart';
-import 'package:concordia_navigation/models/node.dart';
-import 'package:concordia_navigation/services/dijkstra.dart';
+import 'package:concordia_navigation/models/indoor/node.dart';
+import 'package:concordia_navigation/services/indoor/dijkstra.dart';
 import 'package:concordia_navigation/services/search.dart';
 
 class IndoorItinerary {
