+++ lib/services/outdoor/outdoor_itinerary.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:concordia_navigation/models/reachable.dart';
+import 'package:concordia_navigation/models/outdoor/reachable.dart';
 import 'package:concordia_navigation/services/outdoor/directions_service.dart';
 import 'package:concordia_navigation/services/outdoor/location_service.dart';
 import 'package:flutter/material.dart';
