+++ lib/services/search.dart
@@ -1,6 +1,6 @@
 import 'package:concordia_navigation/models/indoor/floor.dart';
 import 'package:concordia_navigation/models/indoor/indoor_location.dart';
-import 'package:concordia_navigation/models/reachable.dart';
+import 'package:concordia_navigation/models/outdoor/reachable.dart';
 import 'package:concordia_navigation/models/uni_location.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
