+++ test/app_widget.dart
@@ -1,5 +1,5 @@
 import 'package:concordia_navigation/main.dart';
-import 'package:concordia_navigation/models/user_location.dart';
+import 'package:concordia_navigation/models/outdoor/user_location.dart';
 import 'package:concordia_navigation/providers/buildings_data.dart';
 import 'package:concordia_navigation/providers/calendar_data.dart';
 import 'package:concordia_navigation/providers/indoor_data.dart';
