+++ test/screens/outdoor_interest_test.dart
@@ -1,9 +1,8 @@
 import 'package:concordia_navigation/models/indoor/indoor_location.dart';
 import 'package:concordia_navigation/models/outdoor/building.dart';
-import 'package:concordia_navigation/models/university.dart';
+import 'package:concordia_navigation/models/outdoor/university.dart';
 import 'package:concordia_navigation/screens/outdoor_interest.dart';
 import 'package:concordia_navigation/services/outdoor/shuttle_service.dart';
-import 'package:concordia_navigation/services/outdoor_poi_list.dart';
 import 'package:concordia_navigation/services/search.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -13,7 +12,6 @@ void main() {
 
   setUp( () async {
     ShuttleService.shuttleSchedule = await ShuttleService.loadJson();
-    OutdoorPOIList.poi = await OutdoorPOIList.loadJson();
 
     List<dynamic> data = await University.loadJson();
     University.concordia = University.fromJson(data);
