+++ test/services/outdoor/indoor_itinerary_test.dart
@@ -1,6 +1,6 @@
 import 'package:concordia_navigation/models/indoor/indoor_location.dart';
-import 'package:concordia_navigation/models/university.dart';
-import 'package:concordia_navigation/services/dijkstra.dart';
+import 'package:concordia_navigation/models/outdoor/university.dart';
+import 'package:concordia_navigation/services/indoor/dijkstra.dart';
 import 'package:concordia_navigation/services/outdoor/indoor_itinerary.dart';
 import 'package:flutter_test/flutter_test.dart';
