+++ test/widgets/map_widget_test.dart
@@ -1,4 +1,3 @@
-import 'package:concordia_navigation/models/university.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
