+++ test/widgets/transportation_mode_widget_test.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'package:concordia_navigation/models/indoor/indoor_location.dart';
 import 'package:concordia_navigation/models/outdoor/building.dart';
-import 'package:concordia_navigation/models/university.dart';
+import 'package:concordia_navigation/models/outdoor/university.dart';
 import 'package:concordia_navigation/providers/map_data.dart';
-import 'package:concordia_navigation/services/dijkstra.dart';
+import 'package:concordia_navigation/services/indoor/dijkstra.dart';
 import 'package:concordia_navigation/services/outdoor/directions_service.dart';
 import 'package:concordia_navigation/services/outdoor/outdoor_itinerary.dart';
 import 'package:concordia_navigation/services/outdoor/shuttle_service.dart';
